- Added coverage for Build and Deploy  - Increased overall stability
